# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Deriv Quant Research & Execution Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /data/backfill:
    post:
      operationId: startBackfill
      tags: [data]
      summary: Start historical data backfill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackfillRequest"
      responses:
        "200":
          description: Backfill started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
  /data/stream/start:
    post:
      operationId: startStream
      tags: [data]
      summary: Start live tick stream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamRequest"
      responses:
        "200":
          description: Stream started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
...
  /data/ticks:
    get:
      operationId: getTicks
      tags: [data]
      summary: Get recent ticks for a symbol
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Tick data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tick"
  /data/candles:
    get:
      operationId: getCandles
      tags: [data]
      summary: Get candles for a symbol and timeframe
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: timeframe
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: Candle data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candle"
  /data/spikes:
    get:
      operationId: getSpikeEvents
      tags: [data]
      summary: Get spike events for a symbol
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Spike events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpikeEvent"
  /models/features/build:
    post:
      operationId: buildFeatures
      tags: [models]
      summary: Build feature set from raw data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureBuildRequest"
      responses:
        "200":
          description: Feature build started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
  /models/train:
    post:
      operationId: trainModel
      tags: [models]
      summary: Train probability model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainModelRequest"
      responses:
        "200":
          description: Training started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
...
  /models/score:
    post:
      operationId: scoreSignal
      tags: [models]
      summary: Score a trade signal using the model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoreRequest"
      responses:
        "200":
          description: Score result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreResult"
...
  /signals/latest:
    get:
      operationId: getLatestSignals
      tags: [signals]
      summary: Get latest trade signals
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Latest signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SignalLog"
...
components:
  schemas:
...
    ActionResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    BackfillRequest:
      type: object
      properties:
        symbol:
          type: string
        days:
          type: integer
          default: 30
      required:
        - symbol
    StreamRequest:
      type: object
      properties:
        symbols:
          type: array
          items:
            type: string
      required:
        - symbols
...
    Tick:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        epochTs:
          type: number
        quote:
          type: number
        createdAt:
          type: string
      required:
        - id
        - symbol
        - epochTs
        - quote
        - createdAt
    Candle:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        timeframe:
          type: string
        openTs:
          type: number
        closeTs:
          type: number
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        tickCount:
          type: integer
      required:
        - id
        - symbol
        - timeframe
        - openTs
        - closeTs
        - open
        - high
        - low
        - close
        - tickCount
    SpikeEvent:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        eventTs:
          type: number
        direction:
          type: string
          enum: [up, down]
        spikeSize:
          type: number
        ticksSincePreviousSpike:
          type: ["integer", "null"]
      required:
        - id
        - symbol
        - eventTs
        - direction
        - spikeSize
        - ticksSincePreviousSpike
    FeatureBuildRequest:
      type: object
      properties:
        symbol:
          type: string
        fromTs:
          type: ["number", "null"]
      required:
        - symbol
    TrainModelRequest:
      type: object
      properties:
        modelName:
          type: string
        symbol:
          type: string
        trainingWindowDays:
          type: integer
          default: 90
      required:
        - modelName
        - symbol
...
    ScoreRequest:
      type: object
      properties:
        symbol:
          type: string
        strategyName:
          type: string
      required:
        - symbol
        - strategyName
    ScoreResult:
      type: object
      properties:
        symbol:
          type: string
        strategyName:
          type: string
        score:
          type: number
        signalType:
          type: string
        confidence:
          type: number
        expectedValue:
          type: number
        regimeCompatible:
          type: boolean
        suggestedDirection:
          type: string
        suggestedSl:
          type: ["number", "null"]
        suggestedTp:
          type: ["number", "null"]
      required:
        - symbol
        - strategyName
        - score
        - signalType
        - confidence
        - expectedValue
        - regimeCompatible
        - suggestedDirection
        - suggestedSl
        - suggestedTp
...
    SignalLog:
      type: object
      properties:
        id:
          type: integer
        ts:
          type: string
        symbol:
          type: string
        strategyName:
          type: string
        score:
          type: number
        expectedValue:
          type: number
        allowedFlag:
          type: boolean
        rejectionReason:
          type: ["string", "null"]
        direction:
          type: ["string", "null"]
        suggestedSl:
          type: ["number", "null"]
        suggestedTp:
          type: ["number", "null"]
        aiVerdict:
          type: ["string", "null"]
        aiReasoning:
          type: ["string", "null"]
        aiConfidenceAdj:
          type: ["number", "null"]
      required:
        - id
        - ts
        - symbol
        - strategyName
        - score
        - expectedValue
        - allowedFlag
        - rejectionReason
        - direction
        - suggestedSl
        - suggestedTp
        - aiVerdict
        - aiReasoning
        - aiConfidenceAdj
...