# lib/api-zod/src/generated/types/scoreResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deriv Quant Research & Execution Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreResult {
  symbol: string;
  strategyName: string;
  score: number;
  signalType: string;
  confidence: number;
  expectedValue: number;
  regimeCompatible: boolean;
  suggestedDirection: string;
  /** @nullable */
  suggestedSl: number | null;
  /** @nullable */
  suggestedTp: number | null;
}