# lib/api-zod/src/generated/types/streamRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deriv Quant Research & Execution Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface StreamRequest {
  symbols: string[];
}